extract MagicMethodMatcher
