+++ src/Rector/Contrib/Nette/Utils/MagicMethodRector.php
@@ -10,8 +10,8 @@ use Rector\Builder\MethodBuilder;
 use Rector\Node\Attribute;
 use Rector\NodeAnalyzer\DocBlockAnalyzer;
 use Rector\Rector\AbstractRector;
+use Rector\Regex\MagicMethodMatcher;
 use Roave\BetterReflection\Reflection\ReflectionClass;
-use Roave\BetterReflection\Reflection\ReflectionProperty;
 
 /**
  * Catches @method annotations at childs of Nette\Object
@@ -23,15 +23,15 @@ use Roave\BetterReflection\Reflection\ReflectionProperty;
  */
 final class MagicMethodRector extends AbstractRector
 {
-    /**
-     * @var string
-     */
-    private const MAGIC_METHODS_PATTERN = '~^
-        [ \t*]*  @method  [ \t]+
-        (?: [^\s(]+  [ \t]+ )?
-        (set|get|is|add)  ([A-Z]\w*)
-        (?: ([ \t]* \()  [ \t]* ([^)$\s]*)  )?
-    ()~mx';
+//    /**
+//     * @var string
+//     */
+//    private const MAGIC_METHODS_PATTERN = '~^
+//        [ \t*]*  @method  [ \t]+
+//        (?: [^\s(]+  [ \t]+ )?
+//        (set|get|is|add)  ([A-Z]\w*)
+//        (?: ([ \t]* \()  [ \t]* ([^)$\s]*)  )?
+//    ()~mx';
 
     /**
      * @var mixed[]
@@ -58,14 +58,21 @@ final class MagicMethodRector extends AbstractRector
      */
     private $currentFileAwareClassReflector;
 
+    /**
+     * @var MagicMethodMatcher
+     */
+    private $magicMethodMatcher;
+
     public function __construct(
         MethodBuilder $methodBuilder,
         DocBlockAnalyzer $docBlockAnalyzer,
-        CurrentFileAwareClassReflector $currentFileAwareClassReflector
+        CurrentFileAwareClassReflector $currentFileAwareClassReflector,
+        MagicMethodMatcher $magicMethodMatcher
     ) {
         $this->methodBuilder = $methodBuilder;
         $this->docBlockAnalyzer = $docBlockAnalyzer;
         $this->currentFileAwareClassReflector = $currentFileAwareClassReflector;
+        $this->magicMethodMatcher = $magicMethodMatcher;
     }
 
     public function isCandidate(Node $node): bool
@@ -96,7 +103,11 @@ final class MagicMethodRector extends AbstractRector
         $currentNamespace = $node->namespacedName->slice(0, -1)
             ->toString();
 
-        $this->magicMethods = $this->matchMagicMethodsInDocComment($currentNamespace, $docComment->getText());
+        $this->magicMethods = $this->magicMethodMatcher->matchInContent(
+            $this->classReflection,
+            $currentNamespace,
+            $docComment->getText()
+        );
 
         return (bool) count($this->magicMethods);
     }
@@ -133,74 +144,5 @@ final class MagicMethodRector extends AbstractRector
         $parentClassName = (string) $classNode->extends->getAttribute(Attribute::RESOLVED_NAME);
 
         return $parentClassName === 'Nette\Object';
-    }
-
-    /**
-     * Mimics https://github.com/nette/utils/blob/v2.3/src/Utils/ObjectMixin.php#L285
-     * only without reflection.
-     *
-     * @todo extract to MagicMethodMatcher service
-     *
-     * @return mixed[]
-     */
-    private function matchMagicMethodsInDocComment(string $currentNamespace, string $text): array
-    {
-        preg_match_all(self::MAGIC_METHODS_PATTERN, $text, $matches, PREG_SET_ORDER);
-
-        $methods = [];
-
-        foreach ($matches as $match) {
-            [$all, $op, $prop, $type] = $match;
-
-            $name = $op . $prop;
-            $prop = strtolower($prop[0]) . substr($prop, 1) . ($op === 'add' ? 's' : '');
-
-            if (! $this->classReflection->hasProperty($prop)) {
-                continue;
-            }
-
-            /** @var ReflectionProperty $propertyReflection */
-            $propertyReflection = $this->classReflection->getProperty($prop);
-
-            if ($propertyReflection && ! $propertyReflection->isStatic()) {
-                $type = $this->resolveType($currentNamespace, $op, $type, $propertyReflection, $match);
-
-                $methods[$name] = [
-                    'propertyType' => $type,
-                    'propertyName' => $prop,
-                    'operation' => $op,
-                ];
-            }
-        }
-
-        return $methods;
-    }
-
-    /**
-     * @param mixed[] $match
-     */
-    private function resolveType(
-        string $currentNamespace,
-        string $op,
-        string $type,
-        ReflectionProperty $propertyReflection,
-        array $match
-    ): ?string {
-        if ($op === 'get' || $op === 'is') {
-            $type = null;
-            $op = 'get';
-        } elseif (! $type && preg_match(
-            '#@var[ \t]+(\S+)' . ($op === 'add' ? '\[\]#' : '#'),
-            $propertyReflection->getDocComment(),
-            $match
-        )) {
-            $type = $match[1];
-        }
-
-        if ($type && $currentNamespace && preg_match('#^[A-Z]\w+(\[|\||\z)#', $type)) {
-            $type = $currentNamespace . '\\' . $type;
-        }
-
-        return $type;
     }
 }
